Fix deep null complex attributes
Complex attributes should have the behaviour that they handle `null`
values recursively. That is:

```javascript
{complex: {foo: null}}
```

is equivalent to:

```javascript
{complex: null}
```

which is itself equivalent to `null`.

This change fixes the `null` check in the attribute compose function,
and checks if the object's leaves are all `null`. If so, then the
attribute itself is removed (if `keepNull == false`).