Move "deep" libraries from lodash
This library imports deep clone and deep equal functionality from
`lodash`.

The `lodash` implementations of this functionality are not particularly
performant, so this change swaps them out for `rfdc` for cloning (~4x
faster) and `fast-deep-equal` for deep equality (~7x faster).

This has the combined effect of making a `Delta.transform` ~3-4x faster.