🐛 Remove deep `null` when composing complex attributes
At the moment, we have a bug when composing our custom "complex"
attributes.

In general, if the `keepNull` flag is set to `false`, the resulting
object should contain *no* `null` values. However, we can currently
achieve this by composing together two complex objects, where one has
a deeply-nested `null` key.

This is because we only check for `null` at the top of the tree when
stripping values.

This change adds a check for all descendants of the attributes, and
strips `null` out there, too.